requirements.txt
google-genai
fastapi
uvicorn
google-cloud-texttospeech